Lab 2 Question 2 Frist commit
You are a traveler and traveling to a country where the currency denominations are
unknown and as you travel, you get to know about the denomination in random order.
You want to make a payment of amount x, in such a way that the number of notes you give
is minimum.
//Assume that the denominations are in such a way that any amount can be paid.
You are a traveler and traveling to a country where the currency denominations are
unknown and as you travel, you get to know about the denomination in random order.
You want to make a payment of amount x, in such a way that the number of notes you give
is minimum.
//Assume that the denominations are in such a way that any amount can be paid.
Input
Take input of all the currency denominations ( random order)
Take input of the amount that you want to pay.

Output
Print the minimum no of notes that you will be using to pay the net amount.