Lab 2 Question 1 Frist Commmit
Question 1
PayMoney. processes thousands of transactions daily amounting to crores of Rupees. They also have a daily target that they must achieve.
 Given a list of transactions done by PayMoney and a daily target,
your task is to determine at which transaction PayMoney achieves the same.
If the target is not achievable, then display the target is not achieved.

Test Case 1 :---

Enter the size of transaction array
3
Enter the values of array
20 12 31
Enter the total no of targets that needs to be achieved
2
Enter the value of target
21
Target achieved after 2 transactions
Enter the value of target
19
Target achieved after 1 transactions
Explanation
Target 1 i.e 21 is achieved after 2 transactions, (20 + 12)
Target 2 i.e 19 is achieved in the 1st transaction itself.

Test Case 2 :--
Enter the size of transaction array
1
Enter the values of array
100
Enter the total no of targets that needs to be achieved
1
Enter the value of target
101
Given target is not achieved
Explanation → Since there is only 1 transaction that is of 100 and the target value is 101,
hence target is not achieved.